shells/loki-grafana-server/src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-sandbox-rest
requests